+++ src/main/java/com/javierhidalgodev/clinicaodontologica/servlets/SvWorkSchedule.java
@@ -1,54 +1,63 @@
 package com.javierhidalgodev.clinicaodontologica.servlets;
 
+import com.javierhidalgodev.clinicaodontologica.dto.user.UserDTO;
 import com.javierhidalgodev.clinicaodontologica.logica.Controller;
-import com.javierhidalgodev.clinicaodontologica.servlets.services.WorkScheduleService;
+import com.javierhidalgodev.clinicaodontologica.services.WorkScheduleService;
+import com.javierhidalgodev.clinicaodontologica.utils.PathUtils;
 import java.io.IOException;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 /**
  *
  * @author Javi
  */
-@WebServlet(name = "SvWorkSchedule", urlPatterns = {"/work-schedule", "/work-schedule/new"})
+@WebServlet(name = "SvWorkSchedule", urlPatterns = {
+    "/work-schedule",
+    "/work-schedule/*"})
 public class SvWorkSchedule extends HttpServlet {
 
-    Controller controller = Controller.getInstance();
     WorkScheduleService workScheduleService = new WorkScheduleService();
 
-    protected void processRequest(HttpServletRequest request, HttpServletResponse response)
+    private boolean doBefore(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
+        HttpSession mySession = request.getSession(false);
+        UserDTO userSession = (UserDTO) mySession.getAttribute("userSession");
+        
+        if(userSession.getRole().equals("user")) {
+            response.sendRedirect(request.getContextPath() + "/index");
+            return false;
+        }
+        
+        return true;
     }
-
+    
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        String servletPath = request.getServletPath();
+//        if(!doBefore(request, response)) return;
+        
+        String pathInfo = request.getPathInfo();
 
-        if (servletPath.contains("new")) {
+        if (pathInfo == null) {
+            workScheduleService.getAllWorkSchedules(request, response);
+            return;
+        }
+
+        if (PathUtils.isPathInfoID(pathInfo) != null) {
+            workScheduleService.getInfo(request, response, PathUtils.isPathInfoID(pathInfo));
+            return;
+        }
+
+        if (PathUtils.isPathInfoNew(pathInfo)) {
             request.getRequestDispatcher("/WEB-INF/views/workScheduleRegisterView.jsp").forward(request, response);
+            return;
         } else {
-            String queryParams = request.getQueryString();
-
-            if (queryParams == null) {
-                // Si no hay query params obtener todos los horarios
-                workScheduleService.getAllWorkSchedules(request, response);
-            } else {
-                // Obtener el param ID y operar condicionalmente
-                String id = request.getParameter("id");
-
-                if (id == null) {
-                    // El parámetro a consultar no existe
-                    // Redirigir al listado de horarios
-                    response.sendRedirect("work-schedule");
-                } else {
-                    // Intentar obtener el horario por id
-                    workScheduleService.editingWorkSchedule(request, response);
-                }
-            }
+            response.sendRedirect(request.getContextPath() + "/work-schedule");
         }
     }
 
@@ -56,26 +65,27 @@ public class SvWorkSchedule extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
 
-        String action = request.getParameter("action");
+        String pathInfo = request.getPathInfo();
 
-        if ("getInfo".equals(action)) {
-            workScheduleService.getInfo(request, response);
-        } else if ("create".equals(action)) {
-            workScheduleService.createWorkSchedule(request, response);
-        } else if ("editing".equals(action)) {
-            workScheduleService.editingWorkSchedule(request, response);
-        } else if ("edit".equals(action)) {
-            workScheduleService.editWorkSchedule(request, response);
-        } else if ("delete".equals(action)) {
-            workScheduleService.deleteWorkSchedule(request, response);
-        } else {
-            response.sendRedirect(request.getContextPath() + "/index");
+        if (pathInfo != null) {
+            Integer workScheduleID = PathUtils.isPathInfoID(pathInfo);
+            String action = request.getParameter("action");
+
+            if (workScheduleID != null && "getInfo".equals(action)) {
+                workScheduleService.getInfo(request, response, workScheduleID);
+            } else {
+                if ("create".equals(action)) {
+                    workScheduleService.createWorkSchedule(request, response);
+                } else if ("delete".equals(action)) {
+                    workScheduleService.deleteWorkSchedule(request, response);
+                } else if ("editing".equals(action)) {
+                    workScheduleService.editingWorkSchedule(request, response, workScheduleID);
+                } else if ("edit".equals(action)) {
+                    workScheduleService.editWorkSchedule(request, response, workScheduleID);
+                } else {
+                    response.sendRedirect(request.getContextPath() + "/index");
+                }
+            }
         }
     }
-
-    @Override
-    public String getServletInfo() {
-        return "Short description";
-    }// </editor-fold>
-
 }
